requirements.txt
regex
groq
python-dotenv
